chore: rework struct of gen tests
